core: move EtcsBrakeParams to osrd-mp
as a dependency of `TrainPhysicsIntegrator`, which is depended on by the
new simulation module

this changes asserts to requires, since kotlin multiplatform doesn't
have asserts for common code for some reason:
https://kotlinlang.org/api/core/kotlin-stdlib/kotlin/assert.html

The original `EtcsBrakeParams.kt` file is renamed to
`EtcsBrakeParamsExt.kt` to fix jacoco duplicate class errors (???)
